Sync 2 (#861)
* Fixes HTTP error codes and errors being swallowed (#826)

* [Tezos] Implement fetching validators based on period type (#830)

* Increase tezos block for getting validators list

* Add fetching validators based on period type

* Rename method to getCurrentValidators

* Improves market error logging (#833)

* Improves market error logging

* Adds further logging

* Fix Compound provider prices (#828)

* respect stored quote currency and fixes fetching compound price

* fix tests

* Update azure-pipelines.yml (#839)

Dev branch will have tests right now

* Add EIP55Checksum for classic market data address (#841)

* Add coingecko checksum (#843)

* Add coingecko checksum

* Fix unit tests

* Postman Tests (#831)

* Add Postman Tests

* Improve domain tests

* Add postman tests for observer

* Update domains data

* Installing newman by postman

* Improve make commands and remove duplicate code

* Add postman tests for market data

* Add tests for batch routes and fix/update swagger

* Specify the host for the Newman make command

* Update makefile comment for Newman

* Filter market data interation data

* Update postman collection

* Limit transaction history to 25 objects (#820)

* Limit transactions to 25

* Add verification before limit

* Remove unused code

* Use Itoa instead FormatInt

* Remove useless TODO message

* Remove unused code for functional tests

* Remove sensitive info (#844)

* Add support to review apps (#846)

* Fix NIMIQ date parser

* Remove integrations tests for Fio testnet

* Remove functional tests for Fio testnet for now

* FIO name lookup: add new chain_code parameter to get_pub_address call (FIO testnet v0.8 -> v0.9). (#853)

* Put back the fio test for integration and functional

* Set minimum amount for cosmos staking (#851)

* Set minimum amount for cosmos

* Fix unit tests

* Remove tests for FIO domain until we have new address

* Split stake api handlers into stake.go (#855)

* Split asset services methods (#856)

* Split asset services methods

* Change method signature

* Improve validators logic using extensions (#857)

* Improve validator logic using extensions

* Fix unit tests

* Fix postman tests for iotex

* add logger for gin

* Add auth parameter for newman make commands

* Remove interation data for maket tests

* add logger to apis
