+++ pkg/tests/postman/transaction_data.json
@@ -81,7 +81,7 @@
     },
     {
         "handler": "iotex",
-        "address": "io1mwekae7qqwlr23220k5n9z3fmjxz72tuchra3m"
+        "address": "io1vg808avg2ydye8djl2axmkc9j0xhzu6vdaw6g5"
     },
     {
         "handler": "icon",
