+++ services/assets/model.go
@@ -1,5 +1,8 @@
 package assets
 
+type AssetValidators []AssetValidator
+type AssetValidatorMap map[string]AssetValidator
+
 type AssetValidator struct {
 	ID          string          `json:"id"`
 	Name        string          `json:"name"`
@@ -15,4 +18,12 @@ type ValidatorPayout struct {
 
 type ValidatorStatus struct {
 	Disabled bool `json:"disabled"`
+}
+
+func (av AssetValidators) toMap() AssetValidatorMap {
+	validators := make(AssetValidatorMap)
+	for _, v := range av {
+		validators[v.ID] = v
+	}
+	return validators
 }
