+++ services/assets/stake_test.go
@@ -39,25 +39,18 @@ var expectedStakeValidator = blockatlas.StakeValidator{
 
 func TestGetImage(t *testing.T) {
 	image := getImage(cosmosCoin, "TGzz8gjYiYRqpfmDwnLxfgPuLVNmpCswVp")
-
 	expected := "https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/cosmos/validators/assets/TGzz8gjYiYRqpfmDwnLxfgPuLVNmpCswVp/logo.png"
-
 	assert.Equal(t, expected, image)
 }
 
 func TestNormalizeValidator(t *testing.T) {
-
 	result := normalizeValidator(validators[0], assets[0], cosmosCoin)
-
 	assert.Equal(t, expectedStakeValidator, result)
 }
 
 func TestNormalizeValidators(t *testing.T) {
-
 	result := normalizeValidators(validators, assets, cosmosCoin)
-
-	expected := []blockatlas.StakeValidator{expectedStakeValidator}
-
+	expected := blockatlas.StakeValidators{expectedStakeValidator}
 	assert.Equal(t, expected, result)
 }
 
@@ -110,30 +103,6 @@ func TestCalcAnnual(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			gotInfo := calculateAnnual(tt.args.annual, tt.args.commission)
 			assert.Equal(t, tt.wanted, gotInfo)
-		})
-	}
-}
-
-func Test_getCoinInfoUrl(t *testing.T) {
-	type args struct {
-		c     coin.Coin
-		token string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{"test Ethereum coin", args{coin.Ethereum(), ""}, AssetsURL + coin.Ethereum().Handle},
-		{"test Ethereum token", args{coin.Ethereum(), "0x0000000000b3F879cb30FE243b4Dfee438691c04"}, AssetsURL + coin.Ethereum().Handle + "/assets/" + "0x0000000000b3F879cb30FE243b4Dfee438691c04"},
-		{"test Binance coin", args{coin.Binance(), ""}, AssetsURL + coin.Binance().Handle},
-		{"test Binance token", args{coin.Binance(), "0x0000000000b3F879cb30FE243b4Dfee438691c04"}, AssetsURL + coin.Binance().Handle + "/assets/" + "0x0000000000b3F879cb30FE243b4Dfee438691c04"},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := getCoinInfoUrl(tt.args.c, tt.args.token); got != tt.want {
-				t.Errorf("getCoinInfoUrl() = %v, want %v", got, tt.want)
-			}
 		})
 	}
 }
